pyproject.toml
[project]
name = "data-engineering-at-converza"
version = "0.1.0"
description = "Multi-layered ETL pipeline for call center analytics with Bronze, Silver, and Gold data layers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.28.0",
    "python-dotenv>=1.0.0",
    "faker>=19.0.0",
    "psycopg2-binary>=2.9.0",
]
